Rename account ID utility functions
sidebar/filter/persona.js used to contain an Angular filter for use in
templates. However following several refactorings it now contains a
simple set of utility functions for parsing account IDs. This renames
the module accordingly and puts it alongside other utility function
modules.
