+++ src/sidebar/components/annotation-header.js
@@ -2,7 +2,7 @@
 
 var annotationMetadata = require('../annotation-metadata');
 var memoize = require('../util/memoize');
-var persona = require('../filter/persona');
+var { isThirdPartyUser, username } = require('../util/account-id');
 
 // @ngInject
 function AnnotationHeaderController(features, groups, settings, serviceUrl) {
@@ -14,8 +14,8 @@ function AnnotationHeaderController(features, groups, settings, serviceUrl) {
 
   this.displayName = () => {
     var userInfo = this.annotation.user_info;
-    var isThirdPartyUser = persona.isThirdPartyUser(this.annotation.user, settings.authDomain);
-    if (features.flagEnabled('client_display_names') || isThirdPartyUser) {
+    var isThirdPartyUser_ = isThirdPartyUser(this.annotation.user, settings.authDomain);
+    if (features.flagEnabled('client_display_names') || isThirdPartyUser_) {
       // userInfo is undefined if the api_render_user_info feature flag is off.
       if (userInfo) {
         // display_name is null if the user doesn't have a display name.
@@ -24,16 +24,16 @@ function AnnotationHeaderController(features, groups, settings, serviceUrl) {
         }
       }
     }
-    return persona.username(this.annotation.user);
+    return username(this.annotation.user);
   };
 
   this.isThirdPartyUser = function () {
-    return persona.isThirdPartyUser(self.annotation.user, settings.authDomain);
+    return isThirdPartyUser(self.annotation.user, settings.authDomain);
   };
 
   this.thirdPartyUsernameLink = function () {
     return settings.usernameUrl ?
-      settings.usernameUrl + persona.username(this.annotation.user):
+      settings.usernameUrl + username(this.annotation.user):
       null;
   };
