+++ src/sidebar/components/annotation.js
@@ -2,7 +2,7 @@
 
 var annotationMetadata = require('../annotation-metadata');
 var events = require('../events');
-var persona = require('../filter/persona');
+var { isThirdPartyUser } = require('../util/account-id');
 
 var isNew = annotationMetadata.isNew;
 var isReply = annotationMetadata.isReply;
@@ -459,7 +459,7 @@ function AnnotationController(
   };
 
   this.isThirdPartyUser = function () {
-    return persona.isThirdPartyUser(self.annotation.user, settings.authDomain);
+    return isThirdPartyUser(self.annotation.user, settings.authDomain);
   };
 
   this.isDeleted = function () {
