+++ src/sidebar/components/hypothesis-app.js
@@ -3,7 +3,7 @@
 var scrollIntoView = require('scroll-into-view');
 
 var events = require('../events');
-var parseAccountID = require('../filter/persona').parseAccountID;
+var { parseAccountID } = require('../util/account-id');
 var scopeTimeout = require('../util/scope-timeout');
 var serviceConfig = require('../service-config');
 var bridgeEvents = require('../../shared/bridge-events');
