+++ src/sidebar/components/login-control.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var bridgeEvents = require('../../shared/bridge-events');
-var persona = require('../filter/persona');
+var { isThirdPartyUser } = require('../util/account-id');
 var serviceConfig = require('../service-config');
 
 module.exports = {
@@ -12,7 +12,7 @@ module.exports = {
     this.serviceUrl = serviceUrl;
 
     this.isThirdPartyUser = function() {
-      return persona.isThirdPartyUser(this.auth.userid, settings.authDomain);
+      return isThirdPartyUser(this.auth.userid, settings.authDomain);
     };
 
     this.shouldShowLogOutButton = function () {
