+++ src/sidebar/util/account-id.js
