+++ src/sidebar/util/test/account-id-test.js
@@ -1,44 +1,44 @@
 'use strict';
 
-var persona = require('../persona');
+var { parseAccountID, username, isThirdPartyUser } = require('../account-id');
 
-describe('persona', function () {
+describe('sidebar.util.account-id', function () {
   var term = 'acct:hacker@example.com';
 
   describe('parseAccountID', function() {
     it('should extract the username and provider', function () {
-      assert.deepEqual(persona.parseAccountID(term), {
+      assert.deepEqual(parseAccountID(term), {
         username: 'hacker',
         provider: 'example.com',
       });
     });
 
     it('should return null if the ID is invalid', function () {
-      assert.equal(persona.parseAccountID('bogus'), null);
+      assert.equal(parseAccountID('bogus'), null);
     });
   });
 
   describe('username', function () {
     it('should return the username from the account ID', function () {
-      assert.equal(persona.username(term), 'hacker');
+      assert.equal(username(term), 'hacker');
     });
 
     it('should return an empty string if the ID is invalid', function () {
-      assert.equal(persona.username('bogus'), '');
+      assert.equal(username('bogus'), '');
     });
   });
 
   describe('isThirdPartyUser', function () {
     it('should return true if user is a third party user', function () {
-      assert.isTrue(persona.isThirdPartyUser('acct:someone@example.com', 'ex.com'));
+      assert.isTrue(isThirdPartyUser('acct:someone@example.com', 'ex.com'));
     });
 
     it('should return false if user is not a third party user', function () {
-      assert.isFalse(persona.isThirdPartyUser('acct:someone@example.com', 'example.com'));
+      assert.isFalse(isThirdPartyUser('acct:someone@example.com', 'example.com'));
     });
 
     it('should return false if the user is invalid', function () {
-      assert.isFalse(persona.isThirdPartyUser('bogus', 'example.com'));
+      assert.isFalse(isThirdPartyUser('bogus', 'example.com'));
     });
   });
 });
